backend/requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
bcrypt==4.0.1
blis==0.7.11
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
confection==0.1.5
cryptography==46.0.6
cymem==2.0.13
dnspython==2.7.0
ecdsa==0.19.2
email-validator==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl#sha256=83872781dc64893d45d9dbe940e05f80df7e7196e169ea29e2e9742fed079549
exceptiongroup==1.3.1
fastapi==0.125.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
langcodes==3.5.1
MarkupSafe==3.0.3
motor==3.7.1
murmurhash==1.0.15
numpy==1.24.3
packaging==26.0
passlib==1.7.4
pathlib_abc==0.1.1
pathy==0.11.0
preshed==3.0.13
pyasn1==0.6.3
pycparser==2.23
pydantic==1.10.26
pymongo==4.16.0
PyPDF2==3.0.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
smart-open==6.4.0
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==0.49.3
thinc==8.1.12
tqdm==4.67.3
typer==0.7.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
wasabi==0.10.1
